// packages/react-select/src/types.d.ts
export interface OptionTypeBase {
  [key: string]: any;
}

export type OptionsType<OptionType extends OptionTypeBase> = ReadonlyArray<
  OptionType
>;

export interface GroupTypeBase<OptionType extends OptionTypeBase> {
  options: OptionsType<OptionType>;
  [key: string]: any;
}

export type GroupedOptionsType<
  OptionType extends OptionTypeBase,
  GroupType extends GroupTypeBase<OptionType> = GroupTypeBase<OptionType>
> = ReadonlyArray<GroupType>;

export type ValueType<
  OptionType extends OptionTypeBase,
  IsMulti extends boolean
> = IsMulti extends true ? OptionsType<OptionType> : OptionType | null;
// ...
export type ClassNamesState = { [key: string]: boolean } | undefined;

export interface CommonProps<
  OptionType extends OptionTypeBase,
  IsMulti extends boolean,
  GroupType extends GroupTypeBase<OptionType> = GroupTypeBase<OptionType>
> {
  clearValue: () => void;
  className?: string;
  cx: (
    state: ClassNamesState | undefined,
    className: string | undefined
  ) => string;
  /**
   * Get the styles of a particular part of the select. Pass in the name of the
   * property as the first argument, and the current props as the second argument.
   * See the `styles` object for the properties available.
   */
  getStyles: (name: string, props: any) => {};
  getValue: () => OptionsType<OptionType>;
  /** Whether the value container currently holds a value. */
  hasValue: boolean;
  /** Set when the value container should hold multiple values */
  isMulti: boolean;
  /** Whether the text is right to left */
  isRtl: boolean;
  options: OptionsType<OptionType>;
  selectOption: (option: OptionType) => void;
  selectProps: SelectProps<OptionType, IsMulti, GroupType>;
  setValue: (
    newValue: ValueType<OptionType, IsMulti>,
    action: SetValueAction,
    option?: OptionType
  ) => void;
  theme: Theme;
}
// ...
export type SetValueAction = 'select-option' | 'deselect-option';
// ...
export interface ThemeSpacing {
  baseUnit: number;
  controlHeight: number;
  menuGutter: number;
}

export interface Theme {
  borderRadius: number;
  colors: { [key: string]: string };
  spacing: ThemeSpacing;
}